jdbcboard/src/main/java/com/example/jdbcboard/repository/JdbcUserRepository.java
package com.example.jdbcboard.repository;

import com.example.jdbcboard.connection.DBConnectionUtil;
import com.example.jdbcboard.domain.User;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Repository;

import java.sql.*;
import java.util.List;

@Slf4j
@Repository
public class JdbcUserRepository implements UserRepository {


    @Override
    public User save(User user) throws SQLException {
        String sql = "INSERT INTO users(nickname, email, password) values (?, ?, ?)";

        Connection connection = null;
        PreparedStatement preparedStatement = null;

        try {
            connection = DBConnectionUtil.getConnection();
            preparedStatement = connection.prepareStatement(sql);
            preparedStatement.setString(1, user.getNickName());
            preparedStatement.setString(2, user.getEmail());
            preparedStatement.setString(3, user.getPassword());
            preparedStatement.executeUpdate();

            return user;
        } catch (SQLException e) {
            log.error("db error", e);
            throw e;
        } finally {
            close(connection, preparedStatement, null);
        }
    }

    private void close(Connection connection, Statement preparedStatement, ResultSet resultSet) {

        if (resultSet != null) {
            try {
                resultSet.close();
            } catch (SQLException e) {
                log.info("error", e);
            }
        }

        if (preparedStatement != null) {
            try {
                preparedStatement.close();
            } catch (SQLException e) {
                log.info("error", e);
            }
        }

        if (connection != null) {
            try {
                connection.close();
            } catch (SQLException e) {
                log.info("error", e);
            }
        }
    }

    @Override
    public User findOneUser(Long userId) {
        return null;
    }

    @Override
    public List<User> findAllUser() {
        return null;
    }
}
